Moved and reconfigured parse_argument and tests.
When creating a unit test that actually used the code in the parse_argument files, it wouldn't work because the code was in an executable, not a library. I moved parse_argument code and header from cmd/dot to lib/common, so it will be included in a library, which can be linked to the tests. The tests are now in a new global tests/ folder, in which all other tests can be placed. This keeps the seperation between code and tests.