Handling -? flag in new command line code.
Removed the -? flag processing from `dotneato_args_initialize` and added it to `gv_process_arguments`. The behavior of the -? flag stays the same, it prints the same usage information. Difference is that `dot -V?` used to print to version, but it prints the usage now.
Made `dotneato_basename` non static so the CmdName can be set prior to the parsing code. This will be changed in the future.