Error message for invalid flags
`gv_process_arguments` now starts with checking if any invalid flags are encountered. If this is the case, it will specify this in an error message and exit. This changes command line behavior. Before `dot -Vrandom` would succeed en print the version, now it will exit and notify the user that it has used invalid flags.