Processing -O in gv_parse_arguments.
Removed the parsing and processing of the -O flag in `config_extra_args` and added processing it to `gv_process_arguments`. The behavior of the flag stay the same: it sets gvc->common.auto_outfile_names to true. The detection of this flag did change. Calling `dot -xO` now sets both Reduce and auto file names to true, this wans't the case before.