Windows: Compiling with /MT instead of /MD
On a release build, compile all projects with /MT instead of /MD. This
means that required system libraries are staticly linked instead, so no
DLL is needed. Before this change, the user was required to install the
SDK, this is no longer the case.

The only exception is gvplugin_gd. This can't be compiled with /MT,
because it's dependency libgd is compiled with /MD.