Processing -q flag in `gv_process_arguments`.
Removed parsing and processing of this flag from `dotneato_args_initialize`. Processing this flag is now done in `gv_process_arguments`. The behavior is almost the same: it calls `agseterr`, but passing 0 will set the error level to `AGWARN` instead of exiting. The function `aggeterr(void)` was added to check if the error was set correctly.