Add GIF output to shapes regression tests
Add comparison function for binary output with the Pillow module
(installation of Pillow is added to the scripts of Appveyor and Travis).

We convert the reference and output images to greyscale, because the
image output has minor differences on different platforms (different
anti-aliasing or slightly different colors). Untill the output is the
same on all platforms, this workaround is required.